.history/requirements_20250302162001.txt
torch
ollama
pymupdf
transformers
sentence-transformers
PyPDF2
flask 
flask-restful 
flasgger
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
Python==3.12.2